@@ -6,7 +6,7 @@ const Head = () => {
       <section className='head'>
         <div className='container flexSB'>
           <div className='logo'>
-            <h1>ACADEMIA</h1>
+            <h1>SPEAK EASY</h1>
             <span>ONLINE EDUCATION & LEARNING</span>
           </div>
 
